move arithmetic tests
