+++ ogma/tests/commands/mod.rs
@@ -1,3 +1,4 @@
 
 
+mod arithmetic;
 mod diagnostics;
